Move common code to root.go
